edit: user api actions refactored into a userRouter file
