+++ backend/package.json
@@ -14,7 +14,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/cookie-parser": "^1.4.2",
+    "@types/morgan": "^1.9.3",
     "connect-mongo": "^4.6.0",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
@@ -22,6 +25,7 @@
     "jest": "^27.4.7",
     "kill-port": "^1.6.1",
     "mongoose": "^6.1.0",
+    "morgan": "^1.10.0",
     "passport": "^0.5.0",
     "passport-discord": "^0.1.4",
     "passport-github2": "^0.1.12",
