+++ backend/src/middleware/isLoggedIn.ts
@@ -1,7 +1,7 @@
 import { Response, NextFunction } from 'express'
 import { IReqAuth } from '../interface'
 
-const auth = async (req: IReqAuth, res: Response, next: NextFunction) => {
+const isLoggedIn = async (req: IReqAuth, res: Response, next: NextFunction) => {
     try {
         if (req.isAuthenticated()) {
             return next()
@@ -14,4 +14,4 @@ const auth = async (req: IReqAuth, res: Response, next: NextFunction) => {
     }
 }
 
-export default auth
+export default isLoggedIn
