+++ backend/src/routes/index.ts
@@ -1,9 +1,9 @@
 import authRouter from './authRouter'
-// import userRouter from './userRouter'
+import userRouter from './userRouter'
 
 const routes = [
     authRouter,
-    // userRouter
+    userRouter
 ]
 
 export default routes
