+++ client/src/hooks/Context.tsx
@@ -8,7 +8,7 @@ export default function Context(props: any) {
     const [userObject, setUserObject] = useState<any>()
 
     useEffect(() => {
-        Axios.get("http://localhost:4000/getuser", { withCredentials: true }).then((res: AxiosResponse) => {
+        Axios.get("http://localhost:4000/api/user/getuser", { withCredentials: true }).then((res: AxiosResponse) => {
             if (res.data) {
                 console.log(res.data)
                 setUserObject(res.data)
